fix(marqo_migration_script): diagnostic transformer to test the timezone storage issue of times as utc but all input as from new york

+++ internal/database/marqo/migration/transformers.go
@@ -10,7 +10,7 @@ import (
 type TransformFunc func(map[string]interface{}) (map[string]interface{}, error)
 
 type TimestampInfo struct {
-	OriginalTimestamp int
+	OriginalTimestamp int64
 	UTCTime string
 	LocalTime string
 	Timezone string
